use direct exchange in event service
Originally one queue was used per topic. This means that for each message only one consumer received a message. Now all messages are sent to a direct exchange for an entity (users, posts etc.) with a routing key for the specific actions. Other services create their own queues and bind them to the exchange.

- Updated IEventservice and EventService
- Updated PostService
- Updated EventSubscriptionService
- Fixed readme
- Added rabbitmq host to all appsettings files
- Used namespace shorthand in files that were using the old syntax